Track in-memory virtual files
While the initial analysis (with diagnostics) runs fine, updating the
VirtualFiles directly causes an exception on the next analysis
complaining that the PSI and VirtualFile are out of sync. Apparently
changing the VirtualFile directly does not automatically cause a
reparse, should we do that manually? Or is there some fancy IntelliJ
machinery for this that I haven't found yet? Updating Document via
FileDocumentManager doesn't seem to do it either.